chore: mv server setup internal
